+++ android/src/main/java/com/cyanchill/missingcore/metadataretriever/utils/Exceptions.kt
@@ -1,0 +1,4 @@
+package com.cyanchill.missingcore.metadataretriever.utils
+
+class RequiredArgumentException(funName: String, argName: String) :
+    IllegalArgumentException("`$argName` must be defined in order to use `$funName`.")