+++ example/src/data/useTracksWithSavedArtwork.ts
@@ -1,13 +1,12 @@
 import {
   MetadataPresets,
   getBulkMetadata,
-  saveArtwork,
+  saveHashedArtwork,
 } from '@missingcore/react-native-metadata-retriever';
 import { useQuery } from '@tanstack/react-query';
 
 import { getAudioFiles } from './getAudioFiles';
-
-import { isFulfilled } from '../utils/promise';
+import { ImageDirectory, getImageDirectory } from '../utils/fs';
 
 export function useTracksWithSavedArtwork(hasPermissions: boolean) {
   return useQuery({
@@ -37,13 +36,43 @@ async function getTracksWithSavedArtwork() {
     audioFiles.map(({ uri }) => uri),
     MetadataPresets.standard
   );
-  const tracksMetadata = await Promise.allSettled(
-    results.results.map(async ({ uri, data }) => {
-      const { id, filename } = assetURIMap[uri]!;
-      const imgUri = await saveArtwork(uri, { compress: 0.8 });
-      return { id, filename, artworkData: imgUri, ...data };
-    })
+
+  const savedHashedImages = new Set(
+    getImageDirectory()
+      .listAsRecords()
+      .map(({ uri, isDirectory }) =>
+        isDirectory ? undefined : uri.split('/').at(-1)?.split('.')[0]
+      )
+      .filter((hash) => hash !== undefined)
   );
+
+  const tracksMetadata: Array<
+    (typeof results)['results'][number]['data'] & {
+      id: string;
+      filename: string;
+      artworkData: string | null;
+    }
+  > = [];
+
+  for (const { uri, data } of results.results) {
+    const { id, filename } = assetURIMap[uri]!;
+    let img: { hash: string; uri: string } | null = null;
+    try {
+      img = await saveHashedArtwork(uri, {
+        saveDirectory: ImageDirectory,
+        knownHashes: Array.from(savedHashedImages),
+        compress: 0.8,
+      });
+    } catch {}
+    if (img?.hash) savedHashedImages.add(img.hash);
+    tracksMetadata.push({
+      id,
+      filename,
+      artworkData: img?.uri || null,
+      ...data,
+    });
+  }
+
   console.log(
     `Got metadata of ${audioFiles.length} tracks in ${(
       (performance.now() - start) /
@@ -54,6 +83,6 @@ async function getTracksWithSavedArtwork() {
 
   return {
     duration: ((performance.now() - start) / 1000).toFixed(4),
-    tracks: tracksMetadata.filter(isFulfilled).map(({ value }) => value),
+    tracks: tracksMetadata,
   };
 }