+++ src/NativeMetadataRetriever.ts
@@ -9,7 +9,10 @@ import { TurboModuleRegistry } from 'react-native';
 */
 
 /** Extra options for when using `getArtwork`. */
-type ArtworkOptions = {
+type MergedArtworkOptions = {
+  /** Specify that we want to return a base64 string representing the image. */
+  base64?: boolean;
+
   /**
    * A value in the range `0.0` - `1.0` specifying the quality of the resulting image.
    * - Defaults to `1`.
@@ -20,17 +23,22 @@ type ArtworkOptions = {
    * - Defaults to `SaveFormat.JPEG`.
    */
   format?: 'jpeg' | 'png' | 'webp';
+
   /** Uri we want to save the artwork to instead of the cache directory. */
   saveUri?: string;
-};
 
-/** Options that can be set to modify the behavior of the package. */
-type ConfigOptions = {
+  // -------------------------------------------------------------------
+  //  Use hashing strategy if the following fields are provided:
+  // -------------------------------------------------------------------
+
+  /** Directory where we want to save the hashed image. The file name will be its hash. */
+  saveDirectory?: string;
   /**
-   * Size of the returned base64 image in MB.
-   * - Defaults to `5`.
+   * An array of known MD5 hashes formatted as a 32-character hexadecimal string
+   * which are stored in `saveDirectory` and is of the same format as what we
+   * pass for the `format` option.
    */
-  maxImageSizeMB?: number | null;
+  knownHashes?: string[];
 };
 
 type DebugInfo = {
@@ -57,14 +65,12 @@ export interface Spec extends TurboModule {
 
   getArtwork(
     uri: string,
-    options: ArtworkOptions & { base64?: boolean }
-  ): Promise<string | null>;
+    options: MergedArtworkOptions
+  ): Promise<{ hash: string; data: string } | null>;
 
   getLyric(uri: string): Promise<string | null>;
 
   getR128Gain(uri: string): Promise<number | null>;
-
-  updateConfigs(options: ConfigOptions): Promise<void>;
 
   /**
    * @deprecated For debugging purposes. Returns an object containing